Added Main Fragment.
